SD/requirements.txt
loguru~=0.6.0
pydantic~=1.10.2
python-dotenv==0.21.0
pymongo~=4.3.3
motor~=3.1.1
grpcio~=1.50.0
grpcio-tools~=1.50.0

uvicorn~=0.19.0
fastapi~=0.87.0